src/weighing_service/requirements.txt
# Websockets
websockets==15.0.1

# Pydantic Settings 
pydantic-settings==2.11.0

# Loguru
loguru==0.7.3